+++ lizard_layers/admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 
 from lizard_layers.models import ValueType
 from lizard_layers.models import AreaValue
+from lizard_layers.models import ParameterType
 
 
 class ValueTypeAdmin(admin.ModelAdmin):
@@ -20,5 +21,14 @@ class AreaValueAdmin(admin.ModelAdmin):
     ]
 
 
+class ParameterTypeAdmin(admin.ModelAdmin):
+    list_display_links = ['value_type']
+    list_display = [
+        'value_type',
+        'parameter',
+    ]
+
+
+admin.site.register(ParameterType, ParameterTypeAdmin)
 admin.site.register(ValueType, ValueTypeAdmin)
 admin.site.register(AreaValue, AreaValueAdmin)